requirements.txt
click==8.0.1
Faker==8.12.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.22
text-unidecode==1.3
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
